crypto/requirements.txt
argcomplete==1.12.3
asgiref==3.3.4
backcall==0.2.0
boto3==1.17.109
botocore==1.20.109
certifi==2021.5.30
cfn-flip==1.2.3
chardet==4.0.0
click==8.0.1
decorator==5.0.9
Django==3.2.4
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
durationpy==0.5
et-xmlfile==1.1.0
future==0.18.2
hjson==3.0.2
idna==2.10
importlib-metadata==4.6.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
jmespath==0.10.0
#kappa==0.6.0
openpyxl==3.0.7
parso==0.7.1
pep517==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
pip-tools==6.2.0
placebo==0.9.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.61.2
traitlets==4.3.3
troposphere==3.0.1
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
#zappa==0.53.0
zipp==3.5.0
